optional/mcp_scaffold/requirements.txt
mcp>=1.12,<2
psycopg[binary]>=3.2,<4
python-dotenv>=1.0,<2
